git-remote-p4: Use a temporal file for the input arguments of p4 print
p4 command is synchronous on the way he consumes the input arguments and
the output of those arguments. As it doesn't consume first all the input
argument and then the output (it seems is dumping the as it parse input
args line by line) we could have problems when we have more than one
file to be "printed". With the temporal file we avoid to have to deal
with two synchronous file descriptors and we just process the output
from "p4" command